data: Fix CanRunHeadless tag in plain gnome .desktop session
In the remote login use case, from a headless greeter, it could happen that
GDM searches for supported sessions in /usr/share/xsessions.
In that case it will find gnome.desktop which had X-GDM-CanRunHeadless=true.

That's a problem because only GNOME wayland sessions support
the remote login use case.

Fix that setting it to false on /usr/share/xsessions, but keep it true
on /usr/share/wayland-sessions.