Split -binary into -leader and -service
Now gnome-session (the shell script) launches the gnome-session-leader,
which talks to systemd over IPC to launch the session. Part of the
session is gnome-session-service (which is what used to be
gnome-session-binary in --systemd-service mode)

There is now no more fallback functionality from systemd startup to the
legacy builtin session. If systemd startup fails, the session doesn't
start.