manager: Expand the definition of "running"
It's not particularly useful to know when the session is exactly in
the RUNNING phase. It's much more useful to know when the session is
actually fully up and running. This happens across three phases:

- INITIALIZATION: Session is still starting up
- APPLICATION: Session _is_ running and functional. Launching autostart
apps doesn't mean that the session itself isn't running
- RUNNING: Trivial
- QUERY_END_SESSION: We're just prompting the user to end the session.
The session itself is still running (and if the user rejects the prompt
we go right back to RUNNING!)
- END_SESSION: Now we're stopping the session, so it's no longer running
- EXIT: Ditto