app: Drop AutostartCondition
This desktop-entry key is:
- Highly GNOME specific, thus not used by normal apps
- Out-of-spec: it should be prefixed with X-GNOME-
- Only used to launch Orca

We can move the code that launches Orca to gnome-settings-daemon, which
already has code for monitoring Orca's enabled gsetting.

We also dropped IsAutostartConditionHandled() completely from dbus.
There was nothing using it when I searched the Debian sources and
GitHub. Which is unsurprising, to say the least