manager: Cleanup phase timeouts
We were only ever using the query_timeout in PHASE_QUERY_END_SESSION,
and the phase_timeout in PHASE_END_SESSION. Since these don't overlap,
we can use the same timeout.

We also adjust the PHASE_END_SESSION timeout back down to 10 seconds,
which is the documented time limit that apps have to respond. I did a
bit of git archaeology: the timeout was originally 10 seconds, but
session _startup_ phases also had a 10 second timeout. Thus, the
timeouts shared one #define. Eventually, the startup timeouts increased
to 30 seconds then to 90 (to deal with live media and slow drives), and
simply nobody noticed that this changed the timeout of PHASE_END_SESSION
to 90 seconds as well.