app: Rework handling of phase
First, we no longer need to store the app's phase because it's always
going to be APPLICATION now. We can assume that.

Second, we now refuse to load apps that set a phase other than
Application. This prevents us from trying to launch services that have
yet to port to systemd unit files