fill: Add Kiosk= option
When Kiosk= is set, then we're looking at a kiosk session. Kiosk
sessions are single-purpose sessions that are designed to run one app.
For instance, GDM uses a kiosk session to run the login screen, and
gnome-initial-setup uses a kiosk session to run gnome-initial-setup.

Kiosk sessions shouldn't autostart any apps, because the session's
entire purpose is to launch one single app. The app the kiosk is
intended to launch should be launched via systemd (and then pulled in
via the gnome-session@session_name.target.d/session.conf drop-in)