leader: Drop --autostart option
This corresponds to a change in GDM, which stopped using this option.

Before, we used autostart for two things. First, to suppress
gnome-session's ability to autolaunch apps in kiosk session. Second, to
ensure that gnome-session autostarts _some_ services (like orca,
spice-vdagent, and other agents)

We can achieve the first goal via the new Kiosk= option, and the second
goal is now handled entirely by systemd

Replaces:
https://gitlab.gnome.org/GNOME/gnome-session/-/merge_requests/41