manager: Switch back to Type=notify
The problem with using BusName is that multiple instances of the manager
might exist in systemd's unit store. For instance: let's say we're
trying to boot into gnome-session@gnome.target, so we have
gnome-session-manager@gnome.service loaded. But let's say that some unit
(in practice, this is IBus) declares that it Conflicts= with
gnome-session@gnome-login.target. This makes systemd load the
gnome-session-manager@gnome-login.service service. But this is a
problem! Since we have two different instances of
gnome-session-manager@.service loaded, we have two different services
trying to claim the same bus name! This is a situation that systemd does
not support.

So let's just go back to Type=notify, which worked. We have an sd_notify
call in the correct spot _anyways_, so we might as well send READY=1
too.

Reverts 236cd838377ff5371028b1f4331ef7d9a51f4c73