leader: Find and set DBUS_SESSION_BUS_ADDRESS
Previously we relied on something to magically set
DBUS_SESSION_BUS_ADDRESS for us before executing us. In practice, this
is a weird legacy codepath in pam_systemd.so, which sets this
environment variable for backwards compatibility with software that
expects it to be set. This is a bit of a layering violation (considering
the socket doesn't even exist at that point) and a hack

In fact, in 2018 [1] systemd tried removing this hack, because the
various dbus client libraries learned to look in $XDG_RUNTIME_DIR/bus
for the session bus. This change was later reverted because some
software misbehaved when $DBUS_SESSION_BUS_ADDRESS is unset

This came up again in the context of gnome-session 49 getting ported to
OpenRC, which doesn't have the DBUS_SESSION_BUS_ADDRESS hacks that
systemd has

This commit cleans this all up. First, we use gdbus's existing logic to
find an appropriate dbus session bus socket even if
DBUS_SESSION_BUS_ADDRESS isn't set in our environment. Second, we set
the environment variable to ensure that it's always set, for backwards
compatibility with software that misbehaves without it.

[1]:
https://github.com/systemd/systemd/commit/2b2b7228bffef626fe8e9f131095995f3d50ee3b